change int to double

+++ lib/data_layer/providers/banking_product_transactions_provider.dart
@@ -22,8 +22,8 @@ class BankingProductTransactionProvider {
     bool forceRefresh = false,
     String? searchString,
     bool? credit,
-    int? amountFrom,
-    int? amountTo,
+    double? amountFrom,
+    double? amountTo,
     DateTime? startDate,
     DateTime? endDate,
   }) async {
